requirements.txt
altgraph==0.17.3
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.11.1
branca==0.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coverage==7.0.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.4.2
Django==3.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.8.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
et-xmlfile==1.1.0
exceptiongroup==1.0.4
Flask==2.2.2
folium==0.14.0
fonttools==4.33.3
gmplot==1.4.1
gunicorn==21.2.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.2
MouseInfo==0.1.3
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
odfpy==1.4.1
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pefile==2023.2.7
Pillow==9.1.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
textblob==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
tweepy==4.10.0
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.5.0
wikipedia==1.4.0
xgboost==1.7.6
xlrd==2.0.1
xlwt==1.3.0
